requirements.txt
Django
djangorestframework
django-filter
django-debug-toolbar
django-eav2
django-mptt
psycopg2-binary
Pillow
requests
django-ckeditor
django-rest-knox
celery[redis]
django-environ
gunicorn
drf-yasg
beautifulsoup4  
humanize
django-admin-rangefilter
djangorestframework-simplejwt